[added] ScrollsOnTransition mixin
When a page’s content is long and you transition
to a new route, the scroll position persists
and its super dumb.

This mixin goes on a route handler that you know
has long content and scrolls the window up when
you leave the page.

Not a bad idea to put this on every route handler

closes #121